+++ routes/api/leonardo.ts
@@ -239,8 +239,8 @@ export const handler = async (
           }
 
           if (isAlone) {
-            htmlContent = `<h2>${i18next.t("leonardo.art", { ns: "translation" })} "<strong>${artResults.name}</strong>" ${i18next.t("leonardo.from_artist", { ns: "translation" })} <strong style="color:${artistResult.color}">${artistResult.last_name}</strong>.</h2>`;
-            htmlContent += `<p class="text-[1rem] leading-none mt-3">${i18next.t("leonardo.discover_arts", { ns: "translation" })} <a href="/art/${artistResult.slug}" class="inline-block"><span class="underline">${i18next.t("leonardo.here", { ns: "translation" })}</span></a>.</p>`;
+            htmlContent = `<h2>${i18next.t("leonardo.art", { ns: "translation" })} "<strong>${artResults.name}</strong>" ${i18next.t("leonardo.from", { ns: "translation" })} <strong style="color:${artistResult.color}">${artistResult.last_name}</strong>.</h2>`;
+            htmlContent += `<p class="text-[1rem] leading-none mt-3">${i18next.t("leonardo.scroll_arts", { ns: "translation" })}</p>`;
           }
         }
         break;